Move Profile to an abstract interface and OciProfile as an implementation
